+++ Src/Attributes/Attributes/DontInjectAttribute.cs
@@ -1,0 +1,7 @@
+﻿namespace FastEndpoints;
+
+/// <summary>
+/// endpoint properties marked with this attribute will disable property injection for that property
+/// </summary>
+[AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
+public class DontInjectAttribute : Attribute { }