+++ Src/Attributes/Attributes/HideFromDocsAttribute.cs
@@ -5,9 +5,3 @@
 /// </summary>
 [AttributeUsage(AttributeTargets.All, AllowMultiple = false)]
 public class HideFromDocsAttribute : Attribute { }
-
-/// <summary>
-/// endpoint properties marked with this attribute will disable property injection for that property
-/// </summary>
-[AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-public class DontInjectAttribute : Attribute { }