+++ Src/Library/Extensions/EndpointExtensions.cs
@@ -1,4 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.DependencyInjection;
 using System.Linq.Expressions;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -10,38 +12,52 @@ internal static class EndpointExtensions
     internal static string ActualName(this Type type)
         => (Nullable.GetUnderlyingType(type) ?? type).Name;
 
-    internal static void Configure(this EndpointDefinition def, BaseEndpoint instance, bool implementsConfigure, object[] epAttribs)
+    internal static void Initialize(this EndpointDefinition def, BaseEndpoint instance, HttpContext? ctx)
     {
-        if (implementsConfigure)
+        instance.Definition = def;
+        instance.HttpContext = ctx!;
+
+        if (def.ImplementsConfigure)
         {
             instance.Configure();
         }
         else
         {
-            foreach (var att in epAttribs)
+            if (def.EpAttributes is not null)
             {
-                switch (att)
+                foreach (var att in def.EpAttributes)
                 {
-                    case HttpAttribute httpAttr:
-                        instance.Verbs(httpAttr.Verb);
-                        def.Routes = httpAttr.Routes;
-                        break;
+                    switch (att)
+                    {
+                        case HttpAttribute httpAttr:
+                            instance.Verbs(httpAttr.Verb);
+                            def.Routes = httpAttr.Routes;
+                            break;
 
-                    case AllowAnonymousAttribute:
-                        def.AllowAnonymous();
-                        break;
+                        case AllowAnonymousAttribute:
+                            def.AllowAnonymous();
+                            break;
 
-                    case AuthorizeAttribute authAttr:
-                        def.Roles(authAttr.Roles?.Split(',') ?? Array.Empty<string>());
-                        def.AuthSchemes(authAttr.AuthenticationSchemes?.Split(',') ?? Array.Empty<string>());
-                        if (authAttr.Policy is not null) def.Policies(new[] { authAttr.Policy });
-                        break;
+                        case AuthorizeAttribute authAttr:
+                            def.Roles(authAttr.Roles?.Split(',') ?? Array.Empty<string>());
+                            def.AuthSchemes(authAttr.AuthenticationSchemes?.Split(',') ?? Array.Empty<string>());
+                            if (authAttr.Policy is not null) def.Policies(new[] { authAttr.Policy });
+                            break;
 
-                    case ThrottleAttribute thrtAttr:
-                        def.Throttle(thrtAttr.HitLimit, thrtAttr.DurationSeconds, thrtAttr.HeaderName);
-                        break;
+                        case ThrottleAttribute thrtAttr:
+                            def.Throttle(thrtAttr.HitLimit, thrtAttr.DurationSeconds, thrtAttr.HeaderName);
+                            break;
+                    }
                 }
             }
+        }
+
+        if (ctx?.RequestServices is not null)
+        {
+            if (def.ValidatorInstance is null && def.ValidatorType is not null)
+                def.ValidatorInstance = ActivatorUtilities.CreateInstance(ctx.RequestServices, def.ValidatorType);
+            if (def.MapperInstance is null && def.MapperType is not null)
+                def.MapperInstance = ActivatorUtilities.CreateInstance(ctx.RequestServices, def.MapperType);
         }
     }