+++ Src/Library/Types.cs
@@ -15,6 +15,9 @@ internal static class Types
     internal static readonly Type EmptyResponse = typeof(EmptyResponse);
     internal static readonly Type EmptyRequest = typeof(EmptyRequest);
     internal static readonly Type EndpointOf2 = typeof(Endpoint<,>);
+    internal static readonly Type EndpointOf3 = typeof(Endpoint<,,>);
+    internal static readonly Type EndpointWithMapperOf2 = typeof(EndpointWithMapper<,>);
+    internal static readonly Type EndpointWithOutRequestOf2 = typeof(EndpointWithoutRequest<,>);
     internal static readonly Type FastEventHandlerOf1 = typeof(FastEventHandler<>);
     internal static readonly Type Http = typeof(Http);
     internal static readonly Type HttpAttribute = typeof(HttpAttribute);